Remove CantorOnS3 in favor of single bucket impl
