+++ cantor-s3/src/main/java/com/salesforce/cantor/s3/ObjectsOnS3.java
@@ -10,7 +10,6 @@ package com.salesforce.cantor.s3;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.AmazonS3Exception;
 import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.salesforce.cantor.s3.utils.S3Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -18,6 +17,7 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Collection;
+import java.util.stream.Collectors;
 
 import static com.salesforce.cantor.common.CommonPreconditions.checkArgument;
 import static com.salesforce.cantor.common.CommonPreconditions.checkCreate;
@@ -28,37 +28,33 @@ import static com.salesforce.cantor.common.ObjectsPreconditions.checkGet;
 import static com.salesforce.cantor.common.ObjectsPreconditions.checkKeys;
 import static com.salesforce.cantor.common.ObjectsPreconditions.checkSize;
 import static com.salesforce.cantor.common.ObjectsPreconditions.checkStore;
-import static com.salesforce.cantor.s3.utils.S3Utils.createBucket;
-import static com.salesforce.cantor.s3.utils.S3Utils.deleteBucket;
 import static com.salesforce.cantor.s3.utils.S3Utils.deleteObject;
+import static com.salesforce.cantor.s3.utils.S3Utils.deleteObjects;
 import static com.salesforce.cantor.s3.utils.S3Utils.getKeys;
 import static com.salesforce.cantor.s3.utils.S3Utils.getObjectBytes;
 import static com.salesforce.cantor.s3.utils.S3Utils.getObjectStream;
+import static com.salesforce.cantor.s3.utils.S3Utils.getSize;
 import static com.salesforce.cantor.s3.utils.S3Utils.putObject;
 
 public class ObjectsOnS3 implements StreamingObjects {
     private static final Logger logger = LoggerFactory.getLogger(ObjectsOnS3.class);
 
-    private static final String bucketNameFormat = "%s-cantor-namespace-%d";
+    // cantor-namespace-<namespace>
+    private static final String namespaceFileFormat = "cantor-namespace-%s";
+    // cantor-object-[<namespace>]-<key>
+    private static final String objectFileFormat = "cantor-object-[%s]-%s";
 
     private final AmazonS3 s3Client;
-    private final String bucketPrefix;
-    private final String bucketNameAllNamespaces;
+    private final String bucketName;
 
-    public ObjectsOnS3(final AmazonS3 s3Client) throws IOException {
-        this(s3Client, "default");
-    }
-
-    public ObjectsOnS3(final AmazonS3 s3Client, final String bucketPrefix) throws IOException {
-        // todo: ensure bucket prefix doesn't include periods (or at least doesn't start/end w/ periods)
+    public ObjectsOnS3(final AmazonS3 s3Client, final String bucketName) throws IOException {
         checkArgument(s3Client != null, "null s3 client");
-        checkString(bucketPrefix, "null/empty bucket prefix");
+        checkString(bucketName, "null/empty bucket name");
         this.s3Client = s3Client;
-        this.bucketPrefix = bucketPrefix;
-        this.bucketNameAllNamespaces = String.format("%s-all-namespaces", bucketPrefix);
+        this.bucketName = bucketName;
         try {
-            if (!this.s3Client.doesBucketExistV2(bucketNameAllNamespaces)) {
-                this.s3Client.createBucket(bucketNameAllNamespaces);
+            if (!this.s3Client.doesBucketExistV2(this.bucketName)) {
+                this.s3Client.createBucket(this.bucketName);
             }
         } catch (final AmazonS3Exception e) {
             logger.warn("exception creating required buckets for objects on s3:", e);
@@ -115,7 +111,7 @@ public class ObjectsOnS3 implements StreamingObjects {
         try {
             return doGet(namespace, key);
         } catch (final AmazonS3Exception e) {
-            logger.warn("exception getting object: " + namespace + "." + key, e);
+            logger.warn("exception getting object: "  + namespace + "." + key, e);
             throw new IOException("exception getting object: " + namespace + "." + key, e);
         }
     }
@@ -124,7 +120,7 @@ public class ObjectsOnS3 implements StreamingObjects {
     public boolean delete(final String namespace, final String key) throws IOException {
         checkDelete(namespace, key);
         try {
-            return doDelete(namespace, key);
+            return deleteObject(this.s3Client, this.bucketName, getObjectName(namespace, key));
         } catch (final AmazonS3Exception e) {
             logger.warn("exception deleting object: " + namespace + "." + key, e);
             throw new IOException("exception deleting object: " + namespace + "." + key, e);
@@ -179,17 +175,26 @@ public class ObjectsOnS3 implements StreamingObjects {
     }
 
     private void doCreate(final String namespace) throws IOException {
-        final String bucket = toBucketName(namespace);
-        createBucket(this.s3Client, bucket);
-        // keep a record of the namespace in the namespaces bucket
-        putObject(this.s3Client, bucketNameAllNamespaces, namespace, new ByteArrayInputStream(bucket.getBytes()), new ObjectMetadata());
+        final String key = getNamespaceKey(namespace);
+        if (this.s3Client.doesObjectExist(this.bucketName, key)) {
+            logger.debug("namespace '{}' already exists; no need to recreate", namespace);
+        }
+
+        logger.info("creating namespace '{}' at '{}.{}'", namespace, this.bucketName, key);
+        // if no exception is thrown, the object was put successfully - ignore response value
+        // TODO: in the future we can use this file to keep track of useful namespace level information like encryption type
+        putObject(this.s3Client, this.bucketName, key, new ByteArrayInputStream(new byte[0]), new ObjectMetadata());
     }
 
-    private void doDrop(final String namespace) throws IOException {
-        final String bucket = toBucketName(namespace);
-        deleteBucket(this.s3Client, bucket);
-        // delete the record in the namespaces bucket
-        deleteObject(this.s3Client, bucketNameAllNamespaces, namespace);
+    private void doDrop(final String namespace) {
+        final String namespaceObjectPrefix = getObjectName(namespace, "");
+        logger.info("dropping namespace '{}'", namespace);
+        logger.debug("deleting all objects with prefix '{}.{}'", this.bucketName, namespaceObjectPrefix);
+        deleteObjects(this.s3Client, this.bucketName, namespaceObjectPrefix);
+
+        final String namespaceKey = getNamespaceKey(namespace);
+        logger.debug("deleting namespace record object '{}.{}'", this.bucketName, namespaceKey);
+        this.s3Client.deleteObject(this.bucketName, namespaceKey);
     }
 
     private void doStore(final String namespace, final String key, final byte[] bytes) throws IOException {
@@ -197,50 +202,57 @@ public class ObjectsOnS3 implements StreamingObjects {
     }
 
     private void doStore(final String namespace, final String key, final InputStream stream, final long length) throws IOException {
-        final String bucket = toBucketName(namespace);
+        final String objectName = getObjectName(namespace, key);
+        if (!this.s3Client.doesObjectExist(this.bucketName, getNamespaceKey(namespace))) {
+            throw new IOException(String.format("namespace '%s' doesn't exist; can't store object with key '%s'", namespace, key));
+        }
+
         final ObjectMetadata metadata = new ObjectMetadata();
         metadata.setContentLength(length);
-        logger.info("storing stream with length={} at '{}.{}'", length, bucket, key);
+        logger.info("storing stream with length={} at '{}.{}'", length, this.bucketName, objectName);
         // if no exception is thrown, the object was put successfully - ignore response value
-        putObject(this.s3Client, bucket, key, stream, metadata);
+        putObject(this.s3Client, this.bucketName, objectName, stream, metadata);
     }
 
     private byte[] doGet(final String namespace, final String key) throws IOException {
-        final String bucket = toBucketName(namespace);
-        logger.debug("retrieving object at '{}.{}'", bucket, key);
-        return getObjectBytes(this.s3Client, bucket, key);
+        final String objectName = getObjectName(namespace, key);
+        logger.debug("retrieving object at '{}.{}'", this.bucketName, objectName);
+        return getObjectBytes(this.s3Client, this.bucketName, objectName);
     }
 
     private InputStream doStream(final String namespace, final String key) throws IOException {
-        final String bucket = toBucketName(namespace);
-        logger.debug("retrieving object as stream at '{}.{}'", bucket, key);
-        return getObjectStream(this.s3Client, bucket, key);
-    }
-
-    private boolean doDelete(final String namespace, final String key) {
-        final String bucket = toBucketName(namespace);
-        logger.debug("deleting object at '{}.{}'", bucket, key);
-        return deleteObject(this.s3Client, bucket, key);
+        final String objectName = getObjectName(namespace, key);
+        if (!this.s3Client.doesObjectExist(this.bucketName, objectName)) {
+            throw new IOException(String.format("couldn't find objectName '%s' for namespace '%s'", objectName, namespace));
+        }
+        return getObjectStream(this.s3Client, this.bucketName, objectName);
     }
 
     private int doSize(final String namespace) {
-        final String bucket = toBucketName(namespace);
-        logger.debug("getting size of bucket '{}'", bucket);
-        return S3Utils.getSize(this.s3Client, bucket, null);
+        return getSize(this.s3Client, this.bucketName, getObjectName(namespace, ""));
     }
 
     private Collection<String> doKeys(final String namespace, final int start, final int count) throws IOException {
-        final String bucket = toBucketName(namespace);
-        logger.debug("getting {} keys for namespace '{}' starting at {}", count, bucket, start);
-        return getKeys(this.s3Client, bucket, null, start, count);
+        final String namespaceObjectPrefix = getObjectName(namespace, "");
+        return getKeys(this.s3Client, this.bucketName, namespaceObjectPrefix, start, count)
+                .stream()
+                .map(objectFile -> objectFile.substring(namespaceObjectPrefix.length()))
+                .collect(Collectors.toList());
     }
 
     private Collection<String> doGetNamespaces() throws IOException {
-        // get all objects in the bucket
-        return getKeys(this.s3Client, this.bucketNameAllNamespaces, null, 0, -1);
+        final String namespacePrefix = getNamespaceKey("");
+        return getKeys(this.s3Client, this.bucketName, namespacePrefix, 0, -1)
+                .stream()
+                .map(namespaceFile -> namespaceFile.substring(namespacePrefix.length()))
+                .collect(Collectors.toList());
     }
 
-    private String toBucketName(final String namespace) {
-        return String.format(bucketNameFormat, this.bucketPrefix, Math.abs(namespace.hashCode()));
+    private String getNamespaceKey(final String namespace) {
+        return String.format(namespaceFileFormat, namespace);
+    }
+
+    private String getObjectName(final String namespace, final String key) {
+        return String.format(objectFileFormat, namespace, key);
     }
 }
